[fix] MBean bug
git-svn-id: https://spf4j.googlecode.com/svn/trunk@273 d408096a-aa36-d13b-a661-48d75a76e993

+++ spf4j-core/src/main/java/org/spf4j/stackmonitor/Sampler.java
@@ -278,7 +278,7 @@ public final class Sampler implements SamplerMBean {
     }
 
     @Override
-    public long getSampleTimeMillis() {
+    public int getSampleTimeMillis() {
         return sampleTimeMillis;
     }
 
@@ -439,7 +439,7 @@ public final class Sampler implements SamplerMBean {
     }
 
     @Override
-    public long getDumpTimeMillis() {
+    public int getDumpTimeMillis() {
         return dumpTimeMillis;
     }
